Initial commit of new DSG text and embedded links
This initial commit of the new text for the Data Submission Guidelines page and its embedded links includes the old text and links for reference, and for comparison of texts once displayed within the UI via the appropriate About page.

Once the new DSG text has been reviewed in situ and corrected/updated, the old text will be deleted, with just the new text being re-committed.